Factor out JSCWasmPlugin
